Struct multipart::client::lazy::LazyError

pub struct LazyError<'a, E> {
    pub field_name: Option<Cow<'a, str>>,
    pub error: E,
    // some fields omitted
}

An error for lazily written multipart requests, including the original error as well as the field which caused the error, if applicable.

Fields

field_name

The field that caused the error. If None, there was a problem opening the stream to write or finalizing the stream.

error

The inner error.