Struct multipart::client::lazy::PreparedFields

pub struct PreparedFields<'d> { /* private fields */ }

The result of Multipart::prepare().

Implements Read, contains the entire request body.

Individual files/streams are dropped as they are read to completion.

Note

The fields in the request may have been reordered to simplify the preparation step. No compliant server implementation will be relying on the specific ordering of fields anyways.

Implementations

impl<'d> PreparedFields<'d>

pub fn to_body<'b>(&'b mut self) -> Body<'b>
where 'd: 'b,

Feature: hyper

Convert self to hyper::client::Body.

impl<'d> PreparedFields<'d>

pub fn content_len(&self) -> Option<u64>

Get the content-length value for this set of fields, if applicable (all fields are sized, i.e. not generic streams).

pub fn boundary(&self) -> &str

Get the boundary that was used to serialize the request.