Module multipart::mock

Mocked types for client-side and server-side APIs.

Structs

ClientRequest
A mock implementation of client::HttpRequest which can spawn an HttpBuffer.

HttpBuffer
A writable buffer which stores the boundary and content-length, if provided.

PanicLogger
Struct that logs if the thread panics before it is dropped.

ServerRequest
A mock implementation of server::HttpRequest that can be read.

StdoutTee
A Write adapter that duplicates all data written to the inner writer as well as stdout.

Functions

log_on_panic
Capture all logging on this thread during PanicLogger's lifetime.