Struct multipart::mock::HttpBuffer

pub struct HttpBuffer {
    pub buf: Vec<u8>,
    pub boundary: String,
    pub content_len: Option<u64>,
    /* private fields */
}

A writable buffer which stores the boundary and content-length, if provided.

Implements client::HttpStream if the client feature is enabled.

Fields
buf: Vec<u8>

The buffer containing the raw bytes.

boundary: String

The multipart boundary.

content_len: Option<u64>

The value of the content-length header, if set.

Implementations

impl HttpBuffer

pub fn new_empty(boundary: String, content_len: Option<u64>) -> HttpBuffer

Create an empty buffer with the given boundary and optional content-length.

pub fn with_buf( buf: Vec<u8>, boundary: String, content_len: Option<u64> ) -> Self

Wrap the given buffer with the given boundary and optional content-length.

pub fn for_server(&self) -> ServerRequest<'_>

Get a ServerRequest wrapping the data in this buffer.