Enum multipart::server::SaveDir

pub enum SaveDir {
    Temp(TempDir),
    Perm(PathBuf),
}

The save directory for Entries. May be temporary (delete-on-drop) or permanent.

Variants

Temp

This directory is temporary and will be deleted, along with its contents, when this wrapper is dropped.

Perm

This directory is permanent and will be left on the filesystem when this wrapper is dropped.

Methods

impl SaveDir

fn as_path(&self) -> &Path

Get the path of this directory, either temporary or permanent.

fn is_temporary(&self) -> bool

Returns true if this is a temporary directory which will be deleted on-drop.

fn into_path(self) -> PathBuf

Unwrap the PathBuf from self; if this is a temporary directory, it will be converted to a permanent one.

fn keep(&mut self)

If this SaveDir is temporary, convert it to permanent. This is a no-op if it already is permanent.

Warning: Potential Data Loss

Even though this will prevent deletion on-drop, the temporary folder on most OSes (where this directory is created by default) can be automatically cleared by the OS at any time, usually on reboot or when free space is low.

It is recommended that you relocate the files from a request which you want to keep to a permanent folder on the filesystem.

fn delete(self) -> Result<()>

Delete this directory and its contents, regardless of its permanence.

Warning: Potential Data Loss

This is very likely irreversible, depending on the OS implementation.

Files deleted programmatically are deleted directly from disk, as compared to most file manager applications which use a staging area from which deleted files can be safely recovered (i.e. Windows' Recycle Bin, OS X's Trash Can, etc.).