Enum multipart::server::SaveResult

pub enum SaveResult {
    Full(Entries),
    Partial(Entries, Error),
    Error(Error),
}

The result of Multipart::save_all().

Variants
Full(Entries)

The operation was a total success. Contained are all entries of the request.

Partial(Entries, Error)

The operation errored partway through. Contained are the entries gathered thus far, as well as the error that ended the process.

Error(Error)

The TempDir for Entries could not be constructed. Contained is the error detailing the problem.

Methods

impl SaveResult

pub fn to_entries(self) -> Option<Entries>

Take the Entries from self, if applicable, and discarding the error, if any.

pub fn to_opt(self) -> (Option<Entries>, Option<Error>)

Decompose self to (Option<Entries>, Option<io::Error>)

pub fn to_result(self) -> Result<Entries>

Map self to an io::Result, discarding the error in the Partial case.