Module multipart::server::iron

Integration with the Iron framework, enabled with the iron feature (optional). Includes a BeforeMiddleware implementation.

Not shown here: implHttpRequestfor iron::Request.

Structs

FileCountLimitError
An error returned from Intercept when the file count limit for a single request was exceeded.

FileSizeLimitError
An error returned from Intercept when the size limit for an individual file is exceeded.

Intercept
A BeforeMiddleware for Iron which will intercept and read-out multipart requests and store the result in the request.

Enums

LimitBehavior
The behavior of Intercept when a file size or count limit is exceeded.

Constants

DEFAULT_FILE_COUNT_LIMIT
The default file count limit for Intercept.

DEFAULT_FILE_SIZE_LIMIT
The default file size limit for Intercept, in bytes.