Enum multipart::server::iron::LimitBehavior

pub enum LimitBehavior {
    ThrowError,
    Continue,
}

The behavior of Intercept when a file size or count limit is exceeded.

Variants

ThrowError

Return an error from the middleware describing the issue.

Continue

Ignore the limit.

In the case of file size limits, the offending file will be truncated in the result.

In the case of file count limits, the request will be completed.