Struct multipart::server::nickel::NickelRequest

pub struct NickelRequest<'mw, 'server, D = ()>where
    'server: 'mw,
    D: 'mw,{
    pub origin: Request<'mw, 'server>,
    pub route_result: Option<RouteResult<'mw, D>>,
    /* private fields */
}

A container for all the request data.

The lifetime 'mw represents the lifetime of various bits of middleware state within nickel. It can vary and get shorter.

The lifetime 'server represents the lifetime of data internal to the server. It is fixed and longer than 'mw.

Fields
origin: Request<'mw, 'server>

the original hyper::server::Request

route_result: Option<RouteResult<'mw, D>>

a HashMap<String, String> holding all params with names and values

Implementations

impl<'mw, 'server, D> Request<'mw, 'server, D>

pub fn from_internal( req: Request<'mw, 'server>, data: &'mw D ) -> Request<'mw, 'server, D>

pub fn param(&self, key: &str) -> Option<&str>

pub fn path_without_query(&self) -> Option<&str>

pub fn server_data(&self) -> &'mw D