Enum multipart::server::save::PartialReason

pub enum PartialReason {
    CountLimit,
    SizeLimit,
    IoError(Error),
}

The reason the save operation quit partway through.

Variants
CountLimit

The count limit for files in the request was hit.

The associated file has not been saved to the filesystem.

SizeLimit

The size limit for an individual file was hit.

The file was partially written to the filesystem.

IoError(Error)

An error occurred during the operation.

Methods

impl PartialReason

fn unwrap_err(self) -> Error

Return io::Error in the IoError case or panic otherwise.

fn expect_err(self, msg: &str) -> Error

Return io::Error in the IoError case or panic with the given message otherwise.