Struct multipart::server::save::BufReader

pub struct BufReader<R, P = StdPolicy> { /* fields omitted */ }

A drop-in replacement for std::io::BufReader with more functionality.

Original method names/signatures and implemented traits are left untouched, making replacement as simple as swapping the import of the type.

By default this type implements the behavior of its std counterpart: it only reads into the buffer when it is empty.

To change this type's behavior, change the policy with .set_policy() using a type from the policy module or your own implementation of ReaderPolicy.

Policies that perform alternating reads and consumes without completely emptying the buffer may benefit from using a ringbuffer via the new_ringbuf() and with_capacity_ringbuf() constructors. Ringbuffers are only available on supported platforms with the slice-deque feature and have some other caveats; see the crate root docs for more details.

Methods

impl<R> BufReader<R, StdPolicy>

pub fn new(inner: R) -> BufReader<R, StdPolicy>

Create a new BufReader wrapping inner, utilizing a buffer of default capacity and the default ReaderPolicy.

pub fn with_capacity(cap: usize, inner: R) -> BufReader<R, StdPolicy>

Create a new BufReader wrapping inner, utilizing a buffer with a capacity of at least cap bytes and the default ReaderPolicy.

The actual capacity of the buffer may vary based on implementation details of the global allocator.

pub fn with_buffer(buf: Buffer, inner: R) -> BufReader<R, StdPolicy>

Wrap inner with an existing Buffer instance and the default ReaderPolicy.

Note

Does not clear the buffer first! If there is data already in the buffer then it will be returned in read() and fill_buf() ahead of any data from inner.

impl<R, P> BufReader<R, P>

pub fn set_policy<P_>(self, policy: P_) -> BufReader<R, P_> where
P_: ReaderPolicy,

Apply a new ReaderPolicy to this BufReader, returning the transformed type.

pub fn policy_mut(&mut self) -> &mut P

Mutate the current ReaderPolicy in-place.

If you want to change the type, use .set_policy().

pub fn policy(&self) -> &P

Inspect the current ReaderPolicy.

pub fn make_room(&mut self)

Move data to the start of the buffer, making room at the end for more reading.

This is a no-op with the *_ringbuf() constructors (requires slice-deque feature).

pub fn reserve(&mut self, additional: usize)

Ensure room in the buffer for at least additional bytes. May not be quite exact due to implementation details of the buffer's allocator.

pub fn buffer(&self) -> &[u8]

Get the section of the buffer containing valid data; may be empty.

Call .consume() to remove bytes from the beginning of this section.

pub fn buf_len(&self) -> usize

Get the current number of bytes available in the buffer.

pub fn capacity(&self) -> usize

Get the total buffer capacity.

pub fn get_ref(&self) -> &R

Get an immutable reference to the underlying reader.

pub fn get_mut(&mut self) -> &mut R

Get a mutable reference to the underlying reader.

Note

Reading directly from the underlying reader is not recommended, as some data has likely already been moved into the buffer.

pub fn into_inner(self) -> R

Consume self and return the inner reader only.

pub fn into_inner_with_buffer(self) -> (R, Buffer)

Consume self and return both the underlying reader and the buffer.

See also: BufReader::unbuffer()

pub fn unbuffer(self) -> Unbuffer<R>

Consume self and return an adapter which implements Read and will empty the buffer before reading directly from the underlying reader.

impl<R, P> BufReader<R, P> where
R: Read,

pub fn read_into_buf(&mut self) -> Result<usize, Error>

Unconditionally perform a read into the buffer.

Does not invoke ReaderPolicy methods.

If the read was successful, returns the number of bytes read.

pub fn boxed<'a>(self) -> BufReader<Box<dyn Read + 'a>, P> where
R: 'a,

Box the inner reader without losing data.