multipart::server::save

Struct PartialEntries

pub struct PartialEntries<M: ReadEntry> {
    pub entries: Entries,
    pub partial: Option<PartialSavedField<M>>,
}

The partial result type for Multipart::save*().

Contains the successfully saved entries as well as the partially saved file that was in the process of being read when the error occurred, if applicable.

Fields
entries: Entries

The entries that were saved successfully.

partial: Option<PartialSavedField<M>>

The field that was in the process of being read. None if the error occurred between entries.

Implementations

impl<M: ReadEntry> PartialEntries<M>

pub fn keep_partial(self) -> Entries

If partial is present and contains a SavedFile then just add it to the Entries instance and return it.

Otherwise, returns self.entries