multipart::server::save

Type Alias EntriesSaveResult

pub type EntriesSaveResult<M> = SaveResult<Entries, PartialEntries<M>>;

Shorthand result for methods that return Entries

Aliased Type

pub enum EntriesSaveResult<M> {
    Full(Entries),
    Partial(PartialEntries<M>, PartialReason),
    Error(Error),
}

Variants

Full(Entries)

The operation was a total success. Contained is the complete result.

Partial(PartialEntries<M>, PartialReason)

The operation quit partway through. Included is the partial result along with the reason.

Error(Error)

An error occurred at the start of the operation, before anything was done.

Implementations

impl<M: ReadEntry> EntriesSaveResult<M>

pub fn into_entries(self) -> Option<Entries>

Take the Entries from self, if applicable, and discarding the error, if any.