Struct multipart::server::FieldHeaders

pub struct FieldHeaders {
    pub name: ArcStr,
    pub filename: Option<String>,
    pub content_type: Option<Mime>,
}

The headers that (may) appear before a multipart/form-data field.

Warning: Values are Client-Provided

Everything in this struct are values from the client and should be considered untrustworthy. This crate makes no effort to validate or sanitize any client inputs.

Fields
name: ArcStr

The field's name from the form.

filename: Option<String>

The filename of this entry, if supplied. This is not guaranteed to match the original file or even to be a valid filename for the current platform.

content_type: Option<Mime>

The MIME type (Content-Type value) of this file, if supplied by the client.

If this is not supplied, the content-type of the field should default to text/plain as per IETF RFC 7578, section 4.4, but this should not be implicitly trusted. This crate makes no attempt to identify or validate the content-type of the actual field data.