Struct multipart::server::MultipartData

pub struct MultipartData<M> { /* private fields */ }

The data of a field in a multipart/form-data request.

You can read it to EOF, or use the save() adaptor to save it to disk/memory.

Implementations

impl<M> MultipartData<M>where M: ReadEntry,

pub fn save(&mut self) -> SaveBuilder<&mut Self>

Get a builder type which can save the field with or without a size limit.

pub fn into_inner(self) -> M

Take the inner Multipart or &mut Multipart

pub fn set_min_buf_size(&mut self, min_buf_size: usize)

Set the minimum buffer size that BufRead::fill_buf(self) will return until the end of the stream is reached. Set this as small as you can tolerate to minimize read() calls (read() wonโt be called again until the buffer is smaller than this).

This value is reset between fields.