Struct multipart::server::MultipartField

pub struct MultipartField<M: ReadEntry> {
    pub name: String,
    pub data: MultipartData<M>,
}

A field in a multipart request. May be either text or a binary stream (file).

Fields
name: String

The field's name from the form

data: MultipartData<M>

The data of the field. Can be text or binary.

Methods

impl<M: ReadEntry> MultipartField<M>

fn next_entry(self) -> ReadEntryResult<M>

Read the next entry in the request.

fn next_entry_inplace(&mut self) -> Result<Option<&mut Self>> where
&'a mut M: ReadEntry,

Update self as the next entry.

Returns Ok(Some(self)) if another entry was read, Ok(None) if the end of the body was reached, and Err(e) for any errors that occur.