Struct multipart::server::MultipartFile

pub struct MultipartFile<'a, B: 'a> {
    // some fields omitted
}

A representation of a file in HTTP multipart/form-data.

Note that the file is not yet saved to the local filesystem; instead, this struct exposes Read and BufRead impls which point to the beginning of the file's contents in the HTTP stream.

You can read it to EOF, or use one of the save_*() methods here to save it to disk.

Methods

impl<'a, B: Read> MultipartFile<'a, B>

fn save_to<W: Write>(&mut self, out: W) -> Result<u64>

Save this file to the given output stream.

If successful, returns the number of bytes written.

Retries when io::Error::kind() == io::ErrorKind::Interrupted.

fn save_to_limited<W: Write>(&mut self, out: W, limit: u64) -> Result<u64>

Save this file to the given output stream, truncated to limit (no more than limit bytes will be written out).

If successful, returns the number of bytes written.

Retries when io::Error::kind() == io::ErrorKind::Interrupted.

fn save_as<P: Into<PathBuf>>(&mut self, path: P) -> Result<SavedFile>

Save this file to path.

Returns the saved file info on success, or any errors otherwise.

Retries when io::Error::kind() == io::ErrorKind::Interrupted.

fn save_in<P: AsRef<Path>>(&mut self, dir: P) -> Result<SavedFile>

Save this file in the directory pointed at by dir, using a random alphanumeric string as the filename.

Any missing directories in the dir path will be created.

Returns the saved file's info on success, or any errors otherwise.

Retries when io::Error::kind() == io::ErrorKind::Interrupted.

fn save_as_limited<P: Into<PathBuf>>(&mut self, path: P, limit: u64) -> Result<SavedFile>

Save this file to path, truncated to limit (no more than limit bytes will be written out).

Any missing directories in the dir path will be created.

Returns the saved file's info on success, or any errors otherwise.

Retries when io::Error::kind() == io::ErrorKind::Interrupted.

fn save_in_limited<P: AsRef<Path>>(&mut self, dir: P, limit: u64) -> Result<SavedFile>

Save this file in the directory pointed at by dir, using a random alphanumeric string as the filename.

Truncates file to limit (no more than limit bytes will be written out).

Any missing directories in the dir path will be created.

Returns the saved file's info on success, or any errors otherwise.

Retries when io::Error::kind() == io::ErrorKind::Interrupted.

fn filename(&self) -> Option<&str>

Get the filename of this entry, if supplied.

Warning

You should treat this value as untrustworthy because it is an arbitrary string provided by the client. You should not blindly append it to a directory path and save the file there, as such behavior could easily be exploited by a malicious client.

fn content_type(&self) -> &Mime

Get the MIME type (Content-Type value) of this file, if supplied by the client, or "applicaton/octet-stream" otherwise.