Crate multipart

Client- and server-side abstractions for HTTP multipart/form-data requests.

Features:

client (default): Enable the client-side abstractions for multipart requests. If the hyper feature is also set, enables integration with the Hyper HTTP client API.

server (default): Enable the server-side abstractions for multipart requests. If the hyper feature is also set, enables integration with the Hyper HTTP server API.

hyper (default): Enable integration with the Hyper HTTP library for client and/or server depending on which other feature flags are set.

iron: Enable integration with the Iron web application framework. See the server::iron module for more information.

tiny_http: Enable integration with the tiny_http crate. See the server::tiny_http module for more information.

nickel_: Enable integration with the Nickel web application framework. See the server::nickel module for more information. Enables the hyper feature.

Modules

client
The client-side abstraction for multipart requests. Enabled with the client feature (on by default).

server
The server-side abstraction for multipart requests. Enabled with the server feature (on by default).

Macros

chain_result!
Chain a series of results together, with or without previous results.