Crate multipart

Client- and server-side abstractions for HTTP multipart/form-data requests.

Features:

This documentation is built with all features enabled.

client: The client-side abstractions for generating multipart requests.

server: The server-side abstractions for parsing multipart requests.

mock: Provides mock implementations of core client and server traits for debugging or non-standard use.

hyper: Integration with the Hyper HTTP library for client and/or server depending on which other feature flags are set.

iron: Integration with the Iron web application framework. See the server::iron module for more information.

nickel (returning in 0.14!): Integration with the Nickel web application framework. See the server::nickel module for more information.

tiny_http: Integration with the tiny_http crate. See the server::tiny_http module for more information.

Note: Work in Progress

I have left a number of Request-for-Comments (RFC) questions on various APIs and other places in the code as there are some cases where I’m not sure what the desirable behavior is.

I have opened an issue as a place to collect responses and discussions for these questions on Github. Please quote the RFC-statement (and/or link to its source line) and provide your feedback there.

Modules

server

The server-side abstraction for multipart requests. Enabled with the server feature.

Macros

chain_result

Chain a series of results together, with or without previous results.