Module multipart::client

The client-side abstraction for multipart requests. Enabled with the client feature (on by default).

Use this when sending POST requests with files to a server.

Modules

hyper
Client-side integration with Hyper. Enabled with the hyper feature (on by default).

Structs

Multipart
The entry point of the client-side multipart API.

SizedRequest
A wrapper around a request object that measures the request body and sets the Content-Length header to its size in bytes.

Traits

HttpRequest
A trait describing an HTTP request that can be used to send multipart data.

HttpStream
A trait describing an open HTTP stream that can be written to.