Module multipart::client

The client-side abstraction for multipart requests. Enabled with the client feature.

Use this when sending POST requests with files to a server.

Modules

lazy
Multipart requests which write out their data in one fell swoop.

Structs

Multipart
The entry point of the client-side multipart API.

SizedRequest
A wrapper around a request object that measures the request body and sets the Content-Length header to its size in bytes.

Traits

HttpRequest
A trait describing an HTTP request that can be used to send multipart data.

HttpStream
A trait describing an open HTTP stream that can be written to.