Module multipart::client::lazy

Multipart requests which write out their data in one fell swoop.

Structs

LazyError

An error for lazily written multipart requests, including the original error as well as the field which caused the error, if applicable.

Multipart

A multipart request which writes all fields at once upon being provided an output stream.

PreparedFields

The result of Multipart::prepare().

Traits

IntoCowPath

Conversion trait necessary for Multipart::add_file() to accept borrowed or owned strings and borrowed or owned paths

Type Aliases

LazyIoError

A LazyError wrapping std::io::Error.

LazyIoResult

Result type for LazyIoError.