Trait multipart::client::HttpStream

pub trait HttpStream: Write {
    type Request: HttpRequest;
    type Response;
    type Error: From<Error> + From<<Self::Request as HttpRequest>::Error>;
    fn finish(self) -> Result<Self::Response, Self::Error>;
}

A trait describing an open HTTP stream that can be written to.

Associated Types

type Request: HttpRequest

The request type that opened this stream.

type Response

The response type that will be returned after the request is completed.

type Error: From<Error> + From<<Self::Request as HttpRequest>::Error>

The error type for this stream. Must be compatible with io::Error as well as Self::Request::Error.

Loading content...
Required methods

fn finish(self) -> Result<Self::Response, Self::Error>

Finalize and close the stream and return the response object, or any error otherwise.

Loading content...