Module multipart::server

The server-side abstraction for multipart requests. Enabled with the server feature (on by default).

Use this when you are implementing an HTTP server and want to to accept, parse, and serve HTTP multipart/form-data requests (file uploads).

See the Multipart struct for more info.

Modules

hyper
Server-side integration with Hyper. Enabled with the hyper feature (on by default).

iron
Integration with the Iron framework, enabled with the iron feature (optional). Includes a BeforeMiddleware implementation.

nickel
Server-side integration with Nickel via the nickel_ feature (optional, enables hyper feature).

tiny_http
Integration with tiny_http with the tiny_http feature (optional).

Structs

Entries
A result of Multipart::save_all().

Multipart
The server-side implementation of multipart/form-data requests.

MultipartField
A field in a multipart request. May be either text or a binary stream (file).

MultipartFile
A representation of a file in HTTP multipart/form-data.

SavedFile
A file saved to the local filesystem from a multipart request.

Enums

MultipartData
The data of a field in a multipart/form-data request.

SaveDir
The save directory for Entries. May be temporary (delete-on-drop) or permanent.

SaveResult
The result of Multipart::save_all().

Traits

HttpRequest
A server-side HTTP request that may or may not be multipart.