Module multipart::server

The server-side abstraction for multipart requests. Enabled with the server feature.

Use this when you are implementing an HTTP server and want to to accept, parse, and serve HTTP multipart/form-data requests (file uploads).

See the Multipart struct for more info.

Re-exports

pub extern crate buf_redux;

pub use self::save::Entries;

pub use self::save::SaveResult;

pub use self::save::SavedField;

Modules

hyper

Server-side integration with Hyper. Enabled with the hyper feature (on by default).

iron

Integration with the Iron framework, enabled with the iron feature (optional). Includes a BeforeMiddleware implementation.

nickel

Support for multipart/form-data bodies in Nickel.

save

Utilities for saving request entries to the filesystem.

tiny_http

Integration with tiny_http with the tiny_http feature (optional).

Structs

FieldHeaders

The headers that (may) appear before a multipart/form-data field.

Multipart

The server-side implementation of multipart/form-data requests.

MultipartData

The data of a field in a multipart/form-data request.

MultipartField

A field in a multipart request with its associated headers and data.

Enums

ReadEntryResult

Ternary result type returned by ReadEntry::next_entry(), Multipart::into_entry() and MultipartField::next_entry().

Traits

HttpRequest

A server-side HTTP request that may or may not be multipart.

ReadEntry

Common trait for Multipart and &mut Multipart