Struct multipart::server::hyper::Switch

pub struct Switch<H, M> {
    // some fields omitted
}

A container that implements hyper::server::Handler which will switch the handler implementation depending on if the incoming request is multipart or not.

Create an instance with new() and pass it to hyper::server::Server::listen() where you would normally pass a Handler instance.

A convenient wrapper for Multipart::from_request().

Methods

impl<H, M> Switch<H, M> where H: Handler, M: MultipartHandler

fn new(normal: H, multipart: M) -> Switch<H, M>

Create a new Switch instance where normal handles normal Hyper requests and multipart handles Multipart requests