Module multipart::server::save

Utilities for saving request entries to the filesystem.

Structs

BufReader
A drop-in replacement for std::io::BufReader with more functionality.

Entries
A result of Multipart::save().

PartialEntries
The partial result type for Multipart::save*().

PartialSavedField
The field that was being read when the save operation quit.

SaveBuilder
A builder for saving a file or files to the local filesystem.

SavedField
A field that has been saved (to memory or disk) from a multipart request.

TempDir
A directory in the filesystem that is automatically deleted when it goes out of scope.

Enums

DataReader
A Read (and BufRead) adapter for SavedData

PartialReason
The reason the save operation quit partway through.

SaveDir
The save directory for Entries. May be temporary (delete-on-drop) or permanent.

SaveResult
The ternary result type used for the SaveBuilder<_> API.

SavedData
A saved field's data container (in memory or on disk)

Type Definitions

EntriesSaveResult
Shorthand result for methods that return Entries

FieldSaveResult
Shorthand result for methods that return FieldDatas.