Struct multipart::server::save::BufReader

pub struct BufReader<R, Rs = IfEmpty, Ms = AtEndLessThan1k> { /* fields omitted */ }

A drop-in replacement for std::io::BufReader with more functionality.

Original method names/signatures and implemented traits are left untouched, making replacement as simple as swapping the import of the type.

Methods

impl<R> BufReader<R, IfEmpty, AtEndLessThan1k>

Important traits for BufReader<R, Rs, Ms>
impl<R, Rs, Ms> Read for BufReader<R, Rs, Ms> where
Ms: MoveStrategy,
R: Read,
Rs: ReadStrategy,

pub fn new(inner: R) -> BufReader<R, IfEmpty, AtEndLessThan1k>

Create a new BufReader wrapping inner, with a buffer of a default capacity and the default strategies.

Important traits for BufReader<R, Rs, Ms>
impl<R, Rs, Ms> Read for BufReader<R, Rs, Ms> where
Ms: MoveStrategy,
R: Read,
Rs: ReadStrategy,

pub fn with_capacity(
cap: usize,
inner: R
) -> BufReader<R, IfEmpty, AtEndLessThan1k>

Create a new BufReader wrapping inner with a capacity of at least cap bytes and the default strategies.

The actual capacity of the buffer may vary based on implementation details of the buffer's allocator.

impl<R, Rs, Ms> BufReader<R, Rs, Ms> where
Ms: MoveStrategy,
Rs: ReadStrategy,

Important traits for BufReader<R, Rs, Ms>
impl<R, Rs, Ms> Read for BufReader<R, Rs, Ms> where
Ms: MoveStrategy,
R: Read,
Rs: ReadStrategy,

pub fn with_strategies(inner: R, rs: Rs, ms: Ms) -> BufReader<R, Rs, Ms>

Create a new BufReader wrapping inner, with a default buffer capacity and with the given ReadStrategy and MoveStrategy.

Important traits for BufReader<R, Rs, Ms>
impl<R, Rs, Ms> Read for BufReader<R, Rs, Ms> where
Ms: MoveStrategy,
R: Read,
Rs: ReadStrategy,

pub fn with_cap_and_strategies(
inner: R,
cap: usize,
rs: Rs,
ms: Ms
) -> BufReader<R, Rs, Ms>

Create a new BufReader wrapping inner, with a buffer capacity of at least cap bytes and the given ReadStrategy and MoveStrategy.

The actual capacity of the buffer may vary based on implementation details of the buffer's allocator.

Important traits for BufReader<R, Rs, Ms>
impl<R, Rs, Ms> Read for BufReader<R, Rs, Ms> where
Ms: MoveStrategy,
R: Read,
Rs: ReadStrategy,

pub fn move_strategy<Ms_>(self, ms: Ms_) -> BufReader<R, Rs, Ms_> where
Ms_: MoveStrategy,

Apply a new MoveStrategy to this BufReader, returning the transformed type.

Important traits for BufReader<R, Rs, Ms>
impl<R, Rs, Ms> Read for BufReader<R, Rs, Ms> where
Ms: MoveStrategy,
R: Read,
Rs: ReadStrategy,

pub fn read_strategy<Rs_>(self, rs: Rs_) -> BufReader<R, Rs_, Ms> where
Rs_: ReadStrategy,

Apply a new ReadStrategy to this BufReader, returning the transformed type.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn move_strategy_mut(&mut self) -> &mut Ms

Accessor for updating the MoveStrategy in-place.

If you want to change the type, use .move_strategy().

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn read_strategy_mut(&mut self) -> &mut Rs

Accessor for updating the ReadStrategy in-place.

If you want to change the type, use .read_strategy().

pub fn make_room(&mut self)

Move data to the start of the buffer, making room at the end for more reading.

pub fn grow(&mut self, additional: usize)

Grow the internal buffer by at least additional bytes. May not be quite exact due to implementation details of the buffer's allocator.

##Note This should not be called frequently as each call will incur a reallocation.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn get_buf(&self) -> &[u8]

Get the section of the buffer containing valid data; may be empty.

Call .consume() to remove bytes from the beginning of this section.

pub fn available(&self) -> usize

Get the current number of bytes available in the buffer.

pub fn capacity(&self) -> usize

Get the total buffer capacity.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_ref(&self) -> &R

Get an immutable reference to the underlying reader.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_mut(&mut self) -> &mut R

Get a mutable reference to the underlying reader.

Note

Reading directly from the underlying reader is not recommended, as some data has likely already been moved into the buffer.

pub fn into_inner(self) -> R

Consume self and return the inner reader only.

pub fn into_inner_with_buf(self) -> (R, Vec<u8>)

Consume self and return both the underlying reader and the buffer, with the data moved to the beginning and the length truncated to contain only valid data.

See also: BufReader::unbuffer()

Important traits for Unbuffer<R>
impl<R> Read for Unbuffer<R> where
R: Read,

pub fn unbuffer(self) -> Unbuffer<R>

Consume self and return an adapter which implements Read and will empty the buffer before reading directly from the underlying reader.

impl<R, Rs, Ms> BufReader<R, Rs, Ms> where
Ms: MoveStrategy,
R: Read,
Rs: ReadStrategy,

pub fn read_into_buf(&mut self) -> Result<usize, Error>

Unconditionally perform a read into the buffer, calling .make_room() if appropriate or necessary, as determined by the implementation.

If the read was successful, returns the number of bytes read.

impl<R, Rs, Ms> BufReader<R, Rs, Ms> where
R: Read,

Important traits for BufReader<R, Rs, Ms>
impl<R, Rs, Ms> Read for BufReader<R, Rs, Ms> where
Ms: MoveStrategy,
R: Read,
Rs: ReadStrategy,

pub fn boxed<'a>(self) -> BufReader<Box<Read + 'a>, Rs, Ms> where
R: 'a,

Box the inner reader without losing data.