Type Alias multipart::server::save::FieldSaveResult

pub type FieldSaveResult = SaveResult<SavedData, SavedData>;

Shorthand result for methods that return FieldDatas.

The MultipartData is not provided here because it is not necessary to return a borrow when the owned version is probably in the same scope. This hopefully saves some headache with the borrow-checker.

Aliased Type

enum FieldSaveResult {
    Full(SavedData),
    Partial(SavedData, PartialReason),
    Error(Error),
}

Variants

Full(SavedData)

The operation was a total success. Contained is the complete result.

Partial(SavedData, PartialReason)

The operation quit partway through. Included is the partial result along with the reason.

Error(Error)

An error occurred at the start of the operation, before anything was done.