Struct multipart::server::MultipartFile

pub struct MultipartFile<'a, R: 'a> {
    // some fields omitted
}

A representation of a file in HTTP multipart/form-data.

Note that the file is not yet saved to the system; instead, this struct exposes Read and BufRead impls which point to the beginning of the file's contents in the HTTP stream.

You can read it to EOF, or use one of the save_*() methods here to save it to disk.

Methods

impl<'a, R: Read> MultipartFile<'a, R>

fn save_as(&mut self, path: &Path) -> Result<u64>

Save this file to path.

Returns the number of bytes written on success, or any errors otherwise.

Retries when io::Error::kind() == io::ErrorKind::Interrupted.

fn save_in(&mut self, dir: &Path) -> Result<PathBuf>

Save this file in the directory pointed at by dir, using self.filename() if present, or a random alphanumeric string otherwise.

Any missing directories in the dir path will be created.

self.filename() is sanitized of all file separators before being appended to dir.

Returns the created file's path on success, or any errors otherwise.

Retries when io::Error::kind() == io::ErrorKind::Interrupted.

fn save(&mut self) -> Result<PathBuf>

Save this file in the directory pointed at by self.save_dir, using self.filename() if present, or a random alphanumeric string otherwise.

Any missing directories in the self.save_dir path will be created.

self.filename() is sanitized of all file separators before being appended to self.save_dir.

Returns the created file's path on success, or any errors otherwise.

Retries when io::Error::kind() == io::ErrorKind::Interrupted.

fn filename(&self) -> Option<&str>

Get the filename of this entry, if supplied.

Warning

You should treat this value as untrustworthy because it is an arbitrary string provided by the client. You should not blindly append it to a directory path and save the file there, as such behavior could easily be exploited by a malicious client.

fn content_type(&self) -> Mime

Get the MIME type (Content-Type value) of this file, if supplied by the client, or "applicaton/octet-stream" otherwise.

fn save_dir(&self) -> &Path

The save directory assigned to this file field by the Multipart instance it was read from.