Trait multipart::server::HttpRequest

pub trait HttpRequest: Read {
    fn multipart_boundary(&self) -> Option<&str>;
}

A server-side HTTP request that may or may not be multipart.

Required Methods

fn multipart_boundary(&self) -> Option<&str>

Get the boundary string of this request if it is a POST request with the Content-Type header set to multipart/form-data.

The boundary string should be supplied as an extra value of the Content-Type header, e.g. Content-Type: multipart/form-data; boundary=myboundary.