Trait multipart::server::HttpRequest

pub trait HttpRequest {
    type Body: Read;

    fn multipart_boundary(&self) -> Option<&str>;
    fn body(self) -> Self::Body;
}

A server-side HTTP request that may or may not be multipart.

May be implemented by mutable references if providing the request or body by-value is undesirable.

Required Associated Types

type Body: Read

The body of this request.

Required Methods

fn multipart_boundary(&self) -> Option<&str>

Get the boundary string of this request if it is a POST request with the Content-Type header set to multipart/form-data.

The boundary string should be supplied as an extra value of the Content-Type header, e.g. Content-Type: multipart/form-data; boundary={boundary}.

fn body(self) -> Self::Body

Return the request body for reading.